Refactored DTTableDatasetNG so that it no longer has mutable parameters.
Previously DTTableDatasetNg kept a mutable Dataset<Row> variable, which was updated when BatchHandler called DTTableDatasetNg.setParagraphDataset().

Now when the data is updated, a new AJAXRequestWatcher with the updated Dataset is created and associated with the AngularObject for the Paragraph.

The AJAXRequestWatcher is now responsible for listening for pagination or search requests from the UI and making the requested operations on its dataset.